refactor(UserType): UserType 병합
