--- src/types/follow/UserType.ts
@@ -1,5 +1,0 @@
-export interface UserType {
-  userId: number;
-  nickname: string;
-  profileImageUrl : string | null;
-}
