+++ src/types/mapData/APIExploreMapType .ts
@@ -1,12 +1,11 @@
-import { UserType } from "./UserType";
-import { KeywordType } from "../keywords/KeywordType";
+import { UserType } from '../UserType';
 
 export interface APIExploreMapType {
   mapId: number;
   imageUrl: string;
   title: string;
-  region : string;
+  region: string;
   description: string;
-  user : UserType;
-  keyword:string[];
+  user: UserType;
+  keyword: string[];
 }
